+++ src/toggl/index.js
@@ -3,7 +3,7 @@ import TogglClient from 'toggl-api'
 import moment from 'moment'
 
 import { SettingsError } from '../errors'
-import { getExtensionSetting, Poller } from '../utils'
+import { getExtensionSetting, logger, Poller } from '../utils'
 
 /**
  * TogglApiClient
@@ -178,18 +178,20 @@ export class TogglApiClient {
   // POLLING
   /**
    * pollCurrentTimeEntry: will poll every x-seconds (min. 3) and get the
-   * currently tracked entry from toggl.com
+   * currently tracked entry from toggl.com. It will also retry 5 times (with an
+   * exponential retry timeout of 2 * interval).
    *
    * @param {function} cb
+   *
+   * polling with retry logic inspired by (thanks):
+   * - https://gitlab.com/snippets/1775781
+   * - https://dev.to/ycmjason/javascript-fetch-retry-upon-failure-3p6g
+   * - https://gist.github.com/briancavalier/842626
    */
   pollCurrentTimeEntry(cb) {
-    const maxRetries = 5
-    let retry = 0
-
     // the actual function that is invoked when a polling cycle starts
-    const pollingFn = async () => {
+    const pollingFn = async (retriesLeft = 5, interval = 1000) => {
       try {
-        // fetch data
         const result = await this.getCurrentTimeEntry()
 
         // send result to callback
@@ -198,16 +200,21 @@ export class TogglApiClient {
         // trigger next polling cycle
         this.poller.poll()
       } catch (error) {
-        // first retry x-times (see maxRetries) before failing
-        if (retry < maxRetries) {
-          retry++
+        logger('error', error)
 
-          // restart polling after an error
-          this.poller.poll()
+        if (retriesLeft) {
+          logger('log', null, `retry polling. ${retriesLeft} tries left...`)
+
+          // wait and retry
+          setTimeout(() => {
+            pollingFn(retriesLeft - 1, interval * 2)
+          }, interval)
+
           return
         }
 
-        cb(error, null)
+        // something went wrong ...
+        cb('Max retries reached', null)
       }
     }
 
@@ -215,7 +222,7 @@ export class TogglApiClient {
     this.poller.onPoll(pollingFn)
 
     // initial start
-    this.poller.poll(true)
+    pollingFn()
   }
 }