+++ src/utils/index.js
@@ -1,4 +1,4 @@
-import { workspace } from 'vscode' // eslint-disable-line
+import { workspace, window } from 'vscode' // eslint-disable-line
 
 import { CONSTANTS } from '../constants'
 
@@ -6,6 +6,15 @@ import Poller from './poller'
 
 export const getExtensionSetting = name => {
   return workspace.getConfiguration(CONSTANTS.prefix)[name]
+}
+
+export const logger = (level, error, msg) => {
+  console[level](`${CONSTANTS.name}: ${error || msg}`)
+
+  if (level === 'error' && msg) {
+    const message = msg || 'An error occured. Please try again.'
+    window.showErrorMessage(message)
+  }
 }
 
 export { Poller }