+++ CHANGELOG.md
@@ -18,6 +18,7 @@ adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 - reduce statusbar width by changing the output text when tracking and when not
 - move from custom method to change statusbar to "internal" (not exposed in
   package.json) vs code command
+- improve polling with better retry-handling
 
 ## 2019/03/06 0.1.0